project/requirements.txt
Django==4.2.7
python-decouple==3.8
PyPDF2==3.0.1
google-generativeai==0.3.2
Pillow==10.1.0